boletin/requirements.txt
Django~=4.1.2

# Django environment variables
# https://django-environ.readthedocs.io/
django-environ

#Postgres Requirement
# sudo apt install python3-dev libpq-dev
#development
psycopg2-binary
#production ready
# psycopg2

djangorestframework~=3.14.0
drf-spectacular
drf-spectacular[sidecar]
pygments